refactor: moving preview wrapper to store
