+++ apps/storefront/app/routes/_store.tsx
@@ -1,7 +1,17 @@
-import { Outlet } from "@remix-run/react";
-import { type LinksFunction } from "@shopify/remix-oxygen";
+import { Outlet, useLoaderData } from "@remix-run/react";
+import {
+  json,
+  type LinksFunction,
+  type LoaderArgs,
+} from "@shopify/remix-oxygen";
+import {
+  isPreviewModeEnabled,
+  Preview,
+  type PreviewData,
+} from "hydrogen-sanity";
 
 import { Layout } from "~/components/global/Layout";
+import { PreviewLoading } from "~/components/global/PreviewLoading";
 import stylesheet from "~/styles/tailwind.css";
 
 export const links: LinksFunction = () => {
@@ -32,10 +42,30 @@ export const links: LinksFunction = () => {
   ];
 };
 
+export async function loader({ context }: LoaderArgs) {
+  const preview: PreviewData | undefined = isPreviewModeEnabled(
+    context.sanity.preview
+  )
+    ? {
+        projectId: context.sanity.preview.projectId,
+        dataset: context.sanity.preview.dataset,
+        token: context.sanity.preview.token,
+      }
+    : undefined;
+
+  return json({
+    preview,
+  });
+}
+
 export default function Store() {
+  const { preview } = useLoaderData<typeof loader>();
+
   return (
-    <Layout>
-      <Outlet />
-    </Layout>
+    <Preview preview={preview} fallback={<PreviewLoading />}>
+      <Layout>
+        <Outlet />
+      </Layout>
+    </Preview>
   );
 }
